+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/common/setup/Items.kt
@@ -7,6 +7,9 @@ import site.siredvin.peripheralium.xplat.LibPlatform
 object Items {
     val PERIPHERALIUM_BLEND = LibPlatform.registerItem("peripheralium_blend") { DescriptiveItem(Item.Properties()) }
     val PERIPHERALIUM_DUST = LibPlatform.registerItem("peripheralium_dust") { DescriptiveItem(Item.Properties()) }
+    val PERIPHERALIUM_UPGRADE_TEMPLATE = LibPlatform.registerItem(
+        "peripheralium_upgrade_template",
+    ) { DescriptiveItem(Item.Properties()) }
 
     fun doSomething() {
     }