+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/data/LibRecipeProvider.kt
@@ -7,6 +7,7 @@ import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.item.crafting.Ingredient
 import site.siredvin.peripheralium.common.setup.Blocks
 import site.siredvin.peripheralium.common.setup.Items
+import site.siredvin.peripheralium.data.blocks.TweakedShapedRecipeBuilder
 import site.siredvin.peripheralium.data.blocks.TweakedShapelessRecipeBuilder
 import site.siredvin.peripheralium.data.blocks.TweakedSmeltingRecipeBuilder
 import site.siredvin.peripheralium.xplat.RecipeIngredients
@@ -31,5 +32,13 @@ class LibRecipeProvider(output: PackOutput) : RecipeProvider(output) {
 
         TweakedSmeltingRecipeBuilder.smelting(Ingredient.of(Items.PERIPHERALIUM_BLEND.get()), Items.PERIPHERALIUM_DUST.get(), 0.7f, 200)
             .save(consumer, ResourceLocation("peripheralium:peripheralium_dust_smelting"))
+
+        TweakedShapedRecipeBuilder.shaped(Items.PERIPHERALIUM_UPGRADE_TEMPLATE.get(), 4)
+            .define('P', Ingredient.of(Items.PERIPHERALIUM_DUST.get()))
+            .define('X', ingredients.xpBottle)
+            .pattern("PPP")
+            .pattern("PXP")
+            .pattern("P P")
+            .save(consumer)
     }
 }