+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/data/blocks/LibTurtleUpgradeDataProvider.kt
@@ -1,9 +1,11 @@
 package site.siredvin.peripheralium.data.blocks
 
+import dan200.computercraft.api.turtle.ITurtleUpgrade
 import dan200.computercraft.api.turtle.TurtleUpgradeDataProvider
 import dan200.computercraft.api.turtle.TurtleUpgradeSerialiser
 import net.minecraft.data.PackOutput
 import net.minecraft.resources.ResourceLocation
+import net.minecraft.world.item.Item
 import site.siredvin.peripheralium.xplat.XplatRegistries
 import java.util.function.Consumer
 import java.util.function.Supplier
@@ -22,6 +24,18 @@ abstract class LibTurtleUpgradeDataProvider(output: PackOutput, serializers: Lis
         if (missedSerializers.isNotEmpty()) {
             throw IllegalArgumentException("Some serializers don't have default items: $missedSerializers")
         }
+    }
+
+    fun <V : ITurtleUpgrade> simpleWithCustomItem(serialiser: TurtleUpgradeSerialiser<V>, item: Item): Upgrade<TurtleUpgradeSerialiser<*>> {
+        return simpleWithCustomItem(XplatRegistries.TURTLE_SERIALIZERS.getKey(serialiser), serialiser, item)
+    }
+
+    fun <V : ITurtleUpgrade> simpleWithCustomItem(serialiser: Supplier<TurtleUpgradeSerialiser<V>>, item: Item): Upgrade<TurtleUpgradeSerialiser<*>> {
+        return simpleWithCustomItem(serialiser.get(), item)
+    }
+
+    fun <V : ITurtleUpgrade, S : Item> simpleWithCustomItem(serialiser: Supplier<TurtleUpgradeSerialiser<V>>, item: Supplier<S>): Upgrade<TurtleUpgradeSerialiser<*>> {
+        return simpleWithCustomItem(serialiser.get(), item.get())
     }
 
     abstract fun registerUpgrades(addUpgrade: Consumer<Upgrade<TurtleUpgradeSerialiser<*>>>)