+++ projects/core/src/main/kotlin/site/siredvin/peripheralium/xplat/RecipeIngredients.kt
@@ -20,4 +20,5 @@ interface RecipeIngredients {
 
     val redstone: Ingredient
     val glowstoneDust: Ingredient
+    val xpBottle: Ingredient
 }