+++ projects/forge/src/main/kotlin/site/siredvin/peripheralium/forge/ForgeIngredients.kt
@@ -1,5 +1,6 @@
 package site.siredvin.peripheralium.forge
 
+import net.minecraft.world.item.Items
 import net.minecraft.world.item.crafting.Ingredient
 import net.minecraftforge.common.Tags
 import site.siredvin.peripheralium.xplat.RecipeIngredients
@@ -9,4 +10,7 @@ object ForgeIngredients : RecipeIngredients {
         get() = Ingredient.of(Tags.Items.DUSTS_REDSTONE)
     override val glowstoneDust: Ingredient
         get() = Ingredient.of(Tags.Items.DUSTS_GLOWSTONE)
+
+    override val xpBottle: Ingredient
+        get() = Ingredient.of(Items.EXPERIENCE_BOTTLE)
 }