+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `1.20` support
+- Peripheralium upgrade template for smithing recipes
 
 ## [0.5.7] - 2023-06-12